fix: address PR #48 review issues
- Move QmdSearchResult type + getResultPath/getResultText helpers to
  shared/types.ts; deduplicate from qmd.ts and search-tool.ts
- Re-resolve binary paths after auto-install in initQmd() so
  ~/.bun/bin/qmd is found on first-time setup
- Use proper SessionMessageEntry type guard instead of unsafe
  intersection cast in session-lifecycle.ts
- Fix timer leak in searchRelevantMemories timeout race by clearing
  the setTimeout when runSearch resolves first
- Remove unused variable in installInstructions()
- Add token-equivalent comments to budget constants
